2016/day20-go/main_test.go
package main

import (
	"math"
	"os"
	"strings"
	"testing"

	"github.com/stretchr/testify/assert"
	"github.com/stretchr/testify/require"
)

func TestFs1Test(t *testing.T) {
	f, err := os.Open("test.txt")
	require.NoError(t, err)
	v, err := fs1(f)
	require.NoError(t, err)
	assert.Equal(t, 3, v)
}

func TestFs1Input(t *testing.T) {
	f, err := os.Open("input.txt")
	require.NoError(t, err)
	v, err := fs1(f)
	require.NoError(t, err)
	assert.Equal(t, 31053880, v)
}

func TestFs2Test(t *testing.T) {
	f, err := os.Open("test.txt")
	require.NoError(t, err)
	v, err := fs2(f, 9)
	require.NoError(t, err)
	assert.Equal(t, 2, v)
}

func TestFs2Unit(t *testing.T) {
	v, err := fs2(strings.NewReader(`2-9
4-8`), 10)
	require.NoError(t, err)
	assert.Equal(t, 3, v)
}

func TestFs2Input(t *testing.T) {
	f, err := os.Open("input.txt")
	require.NoError(t, err)
	v, err := fs2(f, math.MaxUint32)
	require.NoError(t, err)
	assert.Equal(t, 537544587, v)
}
